+++ selfdrive/car/gm/carcontroller.py
@@ -142,7 +142,7 @@ class CarController():
           CS.no_friction_braking]]) + ",")
       
       if (CS.one_pedal_mode_active or CS.coast_one_pedal_mode_active):
-        if not CS.one_pedal_mode_active and CS.is_ev and CS.gear_shifter_ev == 4 and CS.vEgo > 0.05:
+        if not CS.one_pedal_mode_active and CS.gear_shifter_ev == 4 and CS.one_pedal_dl_coasting_enabled and CS.vEgo > 0.05:
           apply_gas = apply_gas * lead_long_gas_lockout_factor + float(P.ZERO_GAS ) * (1. - lead_long_gas_lockout_factor)
         else:
           apply_gas = apply_gas * lead_long_gas_lockout_factor + float(P.MAX_ACC_REGEN) * (1. - lead_long_gas_lockout_factor)